requirements.txt
locket
flask
mastodon.py
numpy
matplotlib
validators